# [SPARK-11088] [SQL] Merges partition values using UnsafeProjection