@marmbrus 2015-10-13
Can we also remove an extra `ConvertToUnsafe` here?  Specifically, is the parquet table scan still claiming to produce safe rows when its really producing unsafe ones now?