@liancheng 2015-10-14
@marmbrus Addressed the `ConvertToUnsafe` issue in PR #9125, since this is mostly a separate issue.