@marmbrus 2015-10-19
LGTM, merging to master.